increased the EPS of Okada to get rid off nodal plane weirdness - cannot be too greedy - even if all variables are double precision, eps should be ~1e-8 and not smaller (not sure why)

+++ block_eval_geokada.c
@@ -79,7 +79,7 @@ void block_eval_geookada(COMP_PRECISION *xl,
       }
     }else{
       // rotate displacements pu back into the observational frame
-      rotate_vec(pu,u,fca,-fsa);
+      rotate_vec(pu,u,(double)fca,(double)-fsa);
     }
   }else{			/* don't return displacements */
     for(i=0;i<3;i++)
